third_party/WebKit/LayoutTests/editing/inserting/insert-text-at-tabspan-002.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
    '<div contenteditable>',
      'a',
      '<span style="white-space:pre">	|		</span>',
      'z',
    '</div>'
  ].join(''),
  'insertText x',
  [
    '<div contenteditable>',
      'a',
      '<span style="white-space:pre">	</span>',
      'x|',
      '<span style="white-space:pre">		</span>',
      'z',
    '</div>'
  ].join('')), 'Insert text in tabspan');
</script>
